backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.5.0
upstash-redis==1.1.0
tensorflow==2.16.1; python_version < "3.13"
tensorflow>=2.21,<2.22; python_version >= "3.13"
scikit-learn==1.4.2; python_version < "3.13"
scikit-learn>=1.7,<1.8; python_version >= "3.13"
pandas==2.2.2; python_version < "3.13"
pandas>=2.3,<2.4; python_version >= "3.13"
numpy==1.26.4; python_version < "3.13"
numpy>=2.3,<2.4; python_version >= "3.13"
joblib==1.4.2
apscheduler==3.10.4
slowapi==0.1.9
transformers>=4.36.0
torch>=2.1.0
safetensors>=0.4.0
